Move login to its own module Login return Result and do not exit on some errors
